+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Phrases shouldn't intercept pointer events unless selectable, so selectable things are selectable.
 - Don't scroll to caret on evaluation update.
 - Prevent undo/redo from going to browser.
+- Correctly restore position after undo/read.
 
 ### Maintenance